Make django apps packages install properly
Most python packages were not installing properly because we did not
define all the packages and package_data that should have been
installed. This issue is not happening when installing in dev mode as a
symlink is created to the actual directory, but when installing using
`pip install package_name`, submodules, templates and static files were
not getting installed.

Fixes #32