Change the default location of atcd's sqlite file.
Defaulting to `/var/lib/atc/atcd.db` will cause `atcd` to fail to start
if `/var/lib/atc` is not created before hand.
This will be an issue when people will install from source and we should
default to a directory that we know exists by default on a system.