run atcui-setup when provisioning a host.
since django 1.7 (which is one of our requirements), syncdb has been
decoupled. `migrate` only sync the db state but does not prompt to
create a superuser.
https://docs.djangoproject.com/en/1.7/ref/django-admin/#migrate-app-label-migrationname
This can be ran without user interaction allowing us to automatically
set up our django app!
This also fixes a bug in atcui-setup that was not properly detecting if
it was run as root