[ui] upgrade to react 0.14 and move out of jsx transformer in browser
JSX is great to quickly code react but till now we have been abusing the
JSXTransformer to do the convertion to js on the fly in the browser.
The way to go is to use npm with a CommonJS module [0], also, with the
deprecation of JSXTransformer [1], moving to babel seems to also need to
use CommonJS features or we need to have everything in a single file.

[0] https://facebook.github.io/react/docs/getting-started.html
[1] https://facebook.github.io/react/blog/2015/06/12/deprecating-jstransform-and-react-tools.html