[atcui] unshape both ipv4 and ipv6 when leaving group
When leaving a group, only one of the 2 IP family (the primary) was unshaped. This would require the user to go to the secondary IP with their browser in order to leave group for the secondary ip.
this change make both primary and secondary leave the group, so typically, when a dual stacked device unshape itself by leaving the group, it will unshape both v4 and v6, the group will then be reaped by atcd.